refactor(config/cli): move to run & packages
