+++ packages/cli/src/index.js
@@ -1,64 +1,18 @@
 const meow = require("meow");
-const { log } = require("./log");
-const { withPkgManager, getPkgManger, PkgMangers } = require("./packages");
-const { ui } = require("./ui");
 const _ = require("lodash");
-const { runCmdFactory, exec } = require("./run");
+const { installCmdBuilder, installDependencies } = require("./run");
+const {
+    getPkgManger,
+    depsPackages,
+    basicPackages,
+    typescriptPackages,
+    filterInstalledDeps,
+    getUserDeps,
+} = require("./packages");
+const { ui } = require("./ui");
+const { log } = require("./log");
 
 const cli = meow(null, {});
-
-const basicPackages = {
-    "@feature-sliced/eslint-config": "latest",
-};
-
-const depsPackages = {
-    "eslint-plugin-boundaries": "^2.8.0",
-    "eslint-plugin-import": "^2.25.4",
-};
-
-const typescriptDeps = {
-    "@typescript-eslint/eslint-plugin": "latest",
-    "@typescript-eslint/parser": "latest",
-    "eslint-import-resolver-typescript": "latest",
-};
-
-function isTypeScriptProject(userDeps) {
-    for (const dep of userDeps) {
-        if (dep.includes("@types/") || dep.includes("typescript")) {
-            return true;
-        }
-    }
-    return false;
-}
-
-function installDependencies(installFn, dependencies, dev = true) {
-    const depsString = Object.keys(dependencies).reduce((result, dep) => {
-        const version = dependencies[dep] && `@${dependencies[dep]}`;
-        return `${result} "${dep + version}"`;
-    }, "");
-
-    const installArgs = `${dev && "-D"}${depsString}`;
-
-    installFn(installArgs);
-}
-
-function filterInstalledDeps(installDeps, existDeps) {
-    const exist = Object.keys(existDeps);
-    return Object.keys(installDeps).reduce(
-        (result, dep) => (exist.includes(dep) ? result : { ...result, [dep]: installDeps[dep] }),
-        {},
-    );
-}
-
-function installCmdBuilder(userPkgManager) {
-    const installCmd = PkgMangers[userPkgManager].install;
-    const userExec = withPkgManager(exec, userPkgManager);
-    return runCmdFactory(installCmd, userExec);
-}
-
-function getUserDeps(cli) {
-    return _.merge(cli.pkg.dependencies, cli.pkg.devDependencies);
-}
 
 function bootstrap({ withTs = true, force = false }) {
     log.info("@feature-sliced/eslint-config/cli");
@@ -79,7 +33,7 @@ function bootstrap({ withTs = true, force = false }) {
     if (withTs) {
         installDependencies(
             runInstall,
-            force ? typescriptDeps : filterInstalledDeps(typescriptDeps, userDeps),
+            force ? typescriptPackages : filterInstalledDeps(typescriptPackages, userDeps),
         );
     }
