+++ packages/cli/src/packages.js
@@ -1,12 +1,49 @@
 const fs = require("fs");
 const path = require("path");
 const { log } = require("./log");
+const _ = require("lodash");
+
+const basicPackages = {
+    "@feature-sliced/eslint-config": "latest",
+};
+
+const depsPackages = {
+    "eslint-plugin-boundaries": "^2.8.0",
+    "eslint-plugin-import": "^2.25.4",
+};
+
+const typescriptPackages = {
+    "@typescript-eslint/eslint-plugin": "latest",
+    "@typescript-eslint/parser": "latest",
+    "eslint-import-resolver-typescript": "latest",
+};
 
 const PkgMangers = {
     npm: { lock: "package-lock.json", install: "install" },
     yarn: { lock: "yarn.lock", install: "add" },
     pnpm: { lock: "pnpm-lock.yaml", install: "install" },
 };
+
+function isTypeScriptProject(userDeps) {
+    for (const dep of userDeps) {
+        if (dep.includes("@types/") || dep.includes("typescript")) {
+            return true;
+        }
+    }
+    return false;
+}
+
+function getUserDeps(cli) {
+    return _.merge(cli.pkg.dependencies, cli.pkg.devDependencies);
+}
+
+function filterInstalledDeps(installDeps, existDeps) {
+    const exist = Object.keys(existDeps);
+    return Object.keys(installDeps).reduce(
+        (result, dep) => (exist.includes(dep) ? result : { ...result, [dep]: installDeps[dep] }),
+        {},
+    );
+}
 
 function getPkgManger() {
     const pkgManagersNames = Object.keys(PkgMangers);
@@ -41,4 +78,14 @@ function withPkgManager(cmdExecutor, pkgManager) {
     };
 }
 
-module.exports = { withPkgManager, getPkgManger, PkgMangers };
+module.exports = {
+    withPkgManager,
+    getPkgManger,
+    PkgMangers,
+    basicPackages,
+    depsPackages,
+    typescriptPackages,
+    getUserDeps,
+    filterInstalledDeps,
+    isTypeScriptProject,
+};
